Added environment for toggling some protocols
