+++ source/core/source/CMakeLists.txt
@@ -1,1 +1,6 @@
+target_sources(stream-triggering
+    PRIVATE
+        "${CMAKE_CURRENT_SOURCE_DIR}/request.cpp"
+)
+
 add_subdirectory(queues)
