+++ source/core/source/request.cpp
@@ -1,0 +1,7 @@
+#include "abstract/request.hpp"
+
+namespace Communication
+{
+bool IPC_PROTOCOL_ENABLED    = true;
+bool CREDIT_PROTOCOL_ENABLED = true;
+}  // namespace Communication
